+++ src/css/components/section_wrap.css
@@ -1,0 +1,11 @@
+
+ /* Базові стилі секцій для телефонів */
+.section-wrap{
+    padding-bottom: 60px;
+}
+
+ /* Базові стилі секцій для планшетів і десктопів */
+ @media screen and (min-width: 768px) {
+ .section-wrap{
+    padding-bottom: 100px;
+}}
