+++ src/css/layout/header.css
@@ -29,65 +29,9 @@
     cursor: pointer;
     }
 
-/* Mobile menu window ================= */
-
-.menu-container {
-  position: fixed;
-  top: 0;
-  left: 0;
-  padding: 124px 110px 180px;
-
-  width: 100%;
-  height: 100%;
-  background-color: var(--accent-color);
-  z-index: 900;
-  /* overflow-y: auto; */
-  transform: translateX(100%);
-  transition: var(--transition-dur-and-func);
-}
-
-.menu-container.is-open {
-  transform: translateX(0);
-}
-
-.mobile-menu-btn-close {
-  position: absolute;
-  top: 16px;
-  right: 16px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 0;
-  width: 24px;
-  height: 24px;
-  border: none;
-  background-color: inherit;
-  cursor: pointer;
-}
-
-.mobile-menu{
-  text-align: center;
-}
-
-
-.mobile-menu-link {
-  font-size: 20px;
-  line-height: calc(24/20);
-  text-decoration: none;
-  color: var(--white-color);
-
-}
-
-.mobile-menu-item:not(:last-child){
-  margin-bottom: 26px;
-}
 
 /* ========Tablet styles=========== */
     @media screen and (min-width: 768px){
-
-      .mobile-menu-btn{
-        display: none;
-    }
 
       .header-box{
        justify-content: center;
