+++ src/css/layout/mobile_menu.css
@@ -1,0 +1,58 @@
+/* Mobile menu window ================= */
+
+.menu-container {
+  background-color: var(--accent-color);
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    z-index: 11;
+    transform: translateX(100%);
+    
+
+    transition: var(--transition-dur-and-func);
+  }
+  
+  
+
+  .menu-container.is-open {
+    transform: translateX(0);
+  }
+  
+  .mobile-menu-btn-close {
+    position: absolute;
+    width: 24px;
+    height: 24px;
+    top: 16px;
+    right: 16px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 0;
+    border: none;
+    background-color: green;
+    cursor: pointer;
+  }
+  
+  .mobile-menu-btn-close-icon{
+    stroke: #ffffff;
+  }
+
+  .mobile-menu{
+    text-align: center;
+  }
+  
+  
+  .mobile-menu-link {
+    font-size: 20px;
+    line-height: calc(24/20);
+    text-decoration: none;
+    color: var(--white-color);
+  
+  }
+  
+  .mobile-menu-item:not(:last-child){
+    margin-bottom: 26px;
+  }
+
