+++ src/index.html
@@ -66,6 +66,10 @@
     <!-- Footer -->
     <load ="./partials/footer.html" />
 
+    <!-- Mobile menu-->
+    <load ="./partials/mobile_menu.html" />
+
+   
     <script type="module" src="/main.js"></script>
 
     <!-- Slick js -->
