+++ src/main.js
@@ -1,2 +1,2 @@
 import './js/slick'
-import './js/mobile-menu'
+import './js/menu'
