+++ src/partials/header.html
@@ -2,7 +2,7 @@
       <div class="container header-box">
         <a href="./index.html" class="logo">
           <svg class="logo-header" width="45" height="35">
-            <use href="#"></use>
+            <use href="../img/svg_sprite.svg#icon-logo"></use>
           </svg>
         </a>
         <nav class="navigation">
