+++ src/partials/mobile_menu.html
@@ -1,0 +1,43 @@
+<div class="menu-container js-menu-container" id="mobile-menu">
+    <button class="mobile-menu-btn-close js-close-menu">
+      <svg class="mobile-menu-btn-close-icon" width="24" height="24">
+        <use href="../img/svg_sprite.svg#icon-x"></use>
+      </svg>
+    </button>
+
+    <ul class="mobile-menu list">
+      <li class="mobile-menu-item">
+        <a href="./index.html" class="mobile-menu-link">Anemone</a>
+      </li>
+      <li class="mobile-menu-item">
+        <a href="" class="mobile-menu-link">About us</a>
+      </li>
+      <li class="mobile-menu-item">
+        <a href="#" class="mobile-menu-link">Store</a>
+      </li>
+      <li class="mobile-menu-item">
+        <a href="#" class="mobile-menu-link">Delivery</a>
+      </li>
+      <li class="mobile-menu-item">
+        <a href="#" class="mobile-menu-link">Blog</a>
+      </li>
+    </ul>
+</div>
+
+<!-- <div class="menu js-menu-container" id="mobile-menu">
+    <button class="menu__button js-close-menu">
+        <svg class="menu__icon">
+            <use href="./images/symbol-defs.svg#icon-close"></use>
+        </svg>
+    </button>
+
+    <nav class="menu__nav">
+        <ul class="menu__nav-list list">
+            <li class="menu__nav-item"><a href="#home" class="menu__nav-link js-close-menu link">Home</a></li>
+            <li class="menu__nav-item"><a href="#products" class="menu__nav-link js-close-menu link">Products</a></li>
+            <li class="menu__nav-item"><a href="#gallery" class="menu__nav-link js-close-menu link">Gallery</a></li>
+            <li class="menu__nav-item"><a href="#feedback" class="menu__nav-link js-close-menu link">Feedback</a></li>
+            <li class="menu__nav-item"><a href="#contacts" class="menu__nav-link js-close-menu link">Contacts</a></li>
+        </ul>
+    </nav>
+</div> -->
